pyproject.toml
[tool.poetry]
name = "ml-training-project"
version = "0.1.0"
description = "A machine learning project for training and evaluating models."
authors = ["Your Name <youremail@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.21"
pandas = "^1.3"
scikit-learn = "^0.24"
torch = "^1.9"
torchvision = "^0.10"
pillow = "^8.2"
pyyaml = "^5.4"
tqdm = "^4.62"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"